Expose JavaMilli into the sbt.io package
Allows for reuse, such as in sbt/sbt#4067.
