+++ io/src/test/scala/sbt/internal/io/JavaMilliSpec.scala
@@ -1,0 +1,8 @@
+package sbt.io
+
+import org.scalatest.FlatSpec
+
+final class JavaMilliSpec extends FlatSpec {
+  "JavaMilli" should "be exposed to the sbt.io package" in
+    assertCompiles("""sbt.io.JavaMilli.getModifiedTime("/tmp")""")
+}
